reorganize components folder
