+++ src/web-app/src/app/page.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@/components/shared/client-react-bootstrap";
+import { Button } from "@/components/client-react-bootstrap";
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
