+++ src/web-app/src/app/zamestnanec/layout.tsx
@@ -1,4 +1,4 @@
-import { Container } from "@/components/shared/client-react-bootstrap";
+import { Container } from "@/components/client-react-bootstrap";
 import NavigationBarEmployee from "./NavigationBar";
 
 /**
