+++ src/web-app/src/components/SignOutButton.tsx
