+++ src/web-app/src/components/errorAlert.tsx
